added Thomas Niepraschk's tag autocomplete feature
The work in this patch is taken from the following GH repository:

  https://github.com/niepi/scuttle-autocomplete

The work is Copyright (c) Thomas Niepraschk 2011.

Niepraschk's work is not done as a branch to the scuttle git repository,
includes a series of unrelated changes, and involves a few mistakes. In
this branch, I've provided a clean-up of his work but very little of my
own.

I have made the following change over what Thomas did in his repository:

- Reverted the move of config.inc.php from config.inc.php.example
- Changed the script code to not include simple load JQuery from
  googleapis.com (introducing some privacy issues) but to load it from
  the copies that Thomas included in the repository instead.
- I have also removed from the history several extra files including the
  full dump of his database (!) which he them removed in a subsequent
  commit.

All credit goes to Thomas Niepraschk for his hard work to build the
functionality. I use it and love it!