fixed missing quotes around mp3
In my system, with JQuery 1.5, the previous two commits caused *every*
link to load the Flash MP3 mplayer, even though none of these were
playable. This patch fixes that issue.

The issue, as far as I can tell, seems to be due to a change in the
version of JQuery. But I have only looked close enough to fix the bug.