increased description size to 1000 characters
The default maximum description in Delicious was raised to 1000
characters instead of 255 several years ago. The current setting in
Scuttle makes it impossible for users to import data from their
Delicious account into Scuttle.

When the limit is increased, the old text-box was a bit too small. This
increases it (in line with what Delicious did when they made the
change).

In this patch, tables.sql is updated, but existing installations will
need to alter their database table my hand. In the absense of a system
for migrations, this seems to be able the best we can do.